split database into three headers
