+++ src/CMakeLists.txt
@@ -11,7 +11,9 @@ target_sources(
   PRIVATE "core/mpi/halfDuplexCommBuffer.cpp"
           "core/primitives/vector.cpp"
           "core/time.cpp"
+          "core/collection.cpp"
           "core/database.cpp"
+          "core/document.cpp"
           "core/dictionary.cpp"
           "core/demangle.cpp"
           "core/tokenList.cpp"
